requirements.txt
aiogram>=3.4.1
openai>=1.33.0
python-dotenv>=1.0.1
requests
replicate
